Rename ast.rs to token.rs
