+++ src/lib.rs
@@ -1,6 +1,6 @@
 #[macro_use] extern crate lalrpop_util;
 
-pub mod ast;
+pub mod token;
 pub mod parser;
 pub mod semantic_analyzer;
 lalrpop_mod!(pub decaf);
